feat: creation class rule
