+++ app/build.gradle
@@ -52,7 +52,7 @@ dependencies {
   testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.3'
 
   // For testing Retrofit
-  implementation("com.squareup.okhttp3:okhttp:4.10.0")
+  implementation 'com.squareup.okhttp3:okhttp:4.10.0'
   testImplementation 'com.squareup.okhttp3:mockwebserver:4.10.0'
 
   androidTestImplementation 'androidx.test.ext:junit:1.1.3'
